Stuff I guess
Needs options that don't fill up the entire time or require timings to be in multiples of eachother

+++ src/competitveroboticscalculator/TimeCalc.java
@@ -4,7 +4,8 @@ package competitveroboticscalculator;
 import java.util.*;
 
 public class TimeCalc {
-    public void calcTime(ArrayList<Operation> operations, int n, int time, ArrayList<Operation> current, ArrayList<ArrayList<Operation>> out){
+    
+    public void calcTime(ArrayList<Operation> operations, int n, double time, ArrayList<Operation> current, ArrayList<ArrayList<Operation>> out){
 	if (time == 0){
             ArrayList<Operation> temp = new ArrayList<Operation>(current);
             out.add(temp);
@@ -14,18 +15,18 @@ public class TimeCalc {
             return;
 	}
         for(int i = n; i < operations.size(); i++){
-        	if(time < operations.get(i).time){
-        		return;
-        	}
-        	if (operations.get(i).count == operations.get(i).max)
-    		{
-        		return;
-    		}else{
+            if(time < operations.get(i).time){
+                return;
+            }
+            if (operations.get(i).count == operations.get(i).max)
+            {
+                return;
+            }else{
     		operations.get(i).count += 1;
             	current.add(operations.get(i));
             	calcTime(operations, i, time - operations.get(i).time, current, out);
             	current.remove(current.size()-1);
-    		}
+            }
         }
     }
     
@@ -33,7 +34,7 @@ public class TimeCalc {
         Collections.sort(ops, new Comparator<Operation>(){
             @Override
             public int compare(Operation o1, Operation o2){
-                return Integer.compare(o2.getTime(), o1.getTime());
+                return Double.compare(o2.getTime(), o1.getTime());
             }
         });
     }
